+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# v1.1
+# v1.1 (not yet released)
 
 Added support for finding faces.
+
+Fixed an issue with symbols in comments being assumed to be variables
+(they may be functions or variables).
 
 # v1.0